Add 'out' module; tools for outputting record sets
This module contains three new classes, only two of which users will
really ever need to interact with directly.

- RecordFileWriter: Manages outputting a batch of files. It's designed
  to break large data sets into multiple files seamlessly.
- Output: Essentially just a data class (but not a dataclass) that
  contains relevant parameters for each data set you want to output.
- Job: The main class people should use to filter an incoming record set
  and output data sets.

The overall goal is to allow running Jobs over generators that feed or
stream records from disk through multiple filters to multiple formatters
and (ultimately) to multiple data sets that get saved on disk. Limits
may be set to prevent files from getting too large and to stop
collecting records once a target number of records has been reached.

This does one and only one pass through the record set; it's designed to
be as efficient (in terms of memory, disk space, and time) as possible.